+++ packages/express/src/index.ts
@@ -2,7 +2,7 @@ export const version = 3;
 export * from './build';
 // @ts-expect-error - FIXME: startDevServer types are not exported
 import { startDevServer as nodeStartDevServer } from '@vercel/node';
-import { findEntrypoint } from './build';
+import { entrypointCallback } from './build';
 import type { ShouldServe, StartDevServer } from '@vercel/build-utils';
 
 export const shouldServe: ShouldServe = async opts => {
@@ -22,11 +22,12 @@ export const shouldServe: ShouldServe = async opts => {
  * need to be relative to the shim's location.
  */
 export const startDevServer: StartDevServer = async opts => {
-  const entrypoint = findEntrypoint(opts.files);
+  const entrypoint = await entrypointCallback(opts);
 
   process.env.EXPERIMENTAL_NODE_TYPESCRIPT_ERRORS = '1';
   return nodeStartDevServer({
     ...opts,
     entrypoint,
+    publicDir: 'public',
   });
 };